Recompute flat_arg_fake_tensors after fakification
Sumamry: This fixes the case when some of the input tensors were
real tensors and fakified in `validate_and_convert_non_fake_tensors`,
but `flat_arg_fake_tensors` would not contain all the inputs
because it was computed before the fakification. We fix this by
recomputing `flat_arg_fake_tensors` after fakification as well.

Reviewers: Chillee, voznesenskym

ghstack-source-id: 00044d316ee2ff9b145e41b77d7829d17e7a2e21
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98769