Make Adam optimizer (almost) work
Currently, use faketensor execution for the first optimizer step,
to create correct-shaped fake tensors for optimizer states, and
avoid taking the init path on the second try.

Then we hackily convert the optimizer states to zero tensors.  This is the
worst part of the current impl, as it fails to trace the real optimizer
init and could be incorrect.

Finally, we can wrap the real optimizer state tensors with fake tensors
to perform a real train step trace.

ghstack-source-id: 1a6707409e3f66de0d2c4bcf599f2434fed7d3e4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98647