[FSDP] Include duplicate parameters and modules when calling named_parameters and named_modules
The default option of `named_parameters` and `named_modules` is to remove the duplicated parameters and modules. However, in FSDP, we need to know what parameters are shared. As a result, setting `remove_duplicate` to False is required in FSDP. Without setting `remove_duplicate` to False, FSDP won't be able to discover shared weights in some cases (e.g., the shared weights are in the same module or there are shared modules).

Differential Revision: [D44897935](https://our.internmc.facebook.com/intern/diff/D44897935/)

ghstack-source-id: 185811798
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98912